chore(filter-utils): adds type to FilterMatchMode
In the type `FilterMatchMode`, we list all the available public methods. This is open for extension, as in order to add a new possible type is a simple as creating a new `public static` method.

We use the `Exclude<T,U>` to remove both the `prototype` and the base `filter` method from the available types.